requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.2
django-simple-captcha==0.5.9
django-stdimage==4.0.1
django-tinymce==2.8.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.7
jdcal==1.4
mysqlclient==1.4.2
odfpy==1.4.0
openpyxl==2.5.12
Pillow==6.0.0
Pillow-PIL==0.1.dev0
progressbar2==3.39.3
PyMySQL==0.9.3
python-utils==2.3.0
pytz==2018.7
PyYAML==3.13
requests==2.20.0
six==1.12.0
slugify==0.0.1
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
